Update create products table

+++ database/migrations/mysql/003_CreateProductsTable.php
@@ -44,7 +44,7 @@ class CreateProductsTable
         $table->id( 'id' );
         $table->string( 'name' );
         $table->text( 'description' );
-        $table->numeric( 'price' )->nullable();
+        $table->int( 'price' )->nullable();
         $table->execute();
     }
 }
